Remove CERES_INCLUDE_DIRS.
The CMake targets file for it says this:

NOTE: There is no equivalent of CERES_INCLUDE_DIRS as the exported
      CMake target already includes the definition of its public
      include directories.

Additionally, including it here is actually harmful since it
somehow causes CMake to generate bogus include data that adds in
the build directory to the final output.  Avoid all of this by
just not including it at all.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>